# Conflicts:
#	aeron-cluster/src/main/java/io/aeron/cluster/client/AeronCluster.java

+++ aeron-system-tests/src/test/java/io/aeron/CounterTest.java
@@ -17,6 +17,11 @@ package io.aeron;
 
 import io.aeron.driver.MediaDriver;
 import io.aeron.driver.ThreadingMode;
+import io.aeron.driver.status.SystemCounterDescriptor;
+import io.aeron.exceptions.AeronException;
+import io.aeron.exceptions.ClientTimeoutException;
+import io.aeron.exceptions.ConductorServiceTimeoutException;
+import io.aeron.exceptions.RegistrationException;
 import io.aeron.status.ReadableCounter;
 import io.aeron.test.InterruptAfter;
 import io.aeron.test.InterruptingTestCallback;
@@ -24,16 +29,29 @@ import io.aeron.test.SystemTestWatcher;
 import io.aeron.test.Tests;
 import io.aeron.test.driver.TestMediaDriver;
 import org.agrona.CloseHelper;
+import org.agrona.collections.MutableBoolean;
+import org.agrona.concurrent.AgentInvoker;
 import org.agrona.concurrent.UnsafeBuffer;
 import org.agrona.concurrent.status.CountersReader;
+import org.hamcrest.CoreMatchers;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.extension.RegisterExtension;
+import org.mockito.Mockito;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
+import java.util.Arrays;
+import java.util.concurrent.ThreadLocalRandom;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicReference;
+
+import static io.aeron.Aeron.NULL_VALUE;
+import static org.hamcrest.CoreMatchers.allOf;
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.MatcherAssert.*;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(InterruptingTestCallback.class)
@@ -42,7 +60,7 @@ class CounterTest
     private static final int COUNTER_TYPE_ID = 1101;
     private static final String COUNTER_LABEL = "counter label";
 
-    private final UnsafeBuffer keyBuffer = new UnsafeBuffer(new byte[8]);
+    private final UnsafeBuffer keyBuffer = new UnsafeBuffer(new byte[64]);
     private final UnsafeBuffer labelBuffer = new UnsafeBuffer(new byte[COUNTER_LABEL.length()]);
 
     private Aeron clientA;
@@ -62,7 +80,9 @@ class CounterTest
         driver = TestMediaDriver.launch(
             new MediaDriver.Context()
                 .errorHandler(Tests::onError)
-                .threadingMode(ThreadingMode.SHARED),
+                .threadingMode(ThreadingMode.SHARED)
+                .clientLivenessTimeoutNs(TimeUnit.SECONDS.toNanos(1))
+                .timerIntervalNs(TimeUnit.MILLISECONDS.toNanos(500)),
             testWatcher);
         testWatcher.dataCollector().add(driver.context().aeronDirectory());
 
@@ -74,7 +94,10 @@ class CounterTest
     void after()
     {
         CloseHelper.closeAll(clientA, clientB, driver);
-        driver.context().deleteDirectory();
+        if (null != driver)
+        {
+            driver.context().deleteDirectory();
+        }
     }
 
     @Test
@@ -197,6 +220,298 @@ class CounterTest
         while (!readableCounter.isClosed())
         {
             Tests.sleep(1, "Counter not closed");
+        }
+    }
+
+    @Test
+    @InterruptAfter(10)
+    void shouldBeAbleToAddStaticCounter()
+    {
+        final AvailableCounterHandler availableCounterHandlerClientA = mock(AvailableCounterHandler.class);
+        final UnavailableCounterHandler unavailableCounterHandlerClientA = mock(UnavailableCounterHandler.class);
+        clientA.addAvailableCounterHandler(availableCounterHandlerClientA);
+        clientA.addUnavailableCounterHandler(unavailableCounterHandlerClientA);
+
+        final AvailableCounterHandler availableCounterHandlerClientB = mock(AvailableCounterHandler.class);
+        final UnavailableCounterHandler unavailableCounterHandlerClientB = mock(UnavailableCounterHandler.class);
+        clientB.addAvailableCounterHandler(availableCounterHandlerClientB);
+        clientB.addUnavailableCounterHandler(unavailableCounterHandlerClientB);
+
+        final Counter counter1 = clientA.addStaticCounter(
+            COUNTER_TYPE_ID,
+            keyBuffer,
+            0,
+            keyBuffer.capacity(),
+            labelBuffer,
+            0,
+            COUNTER_LABEL.length(),
+            100);
+
+        assertFalse(counter1.isClosed());
+        assertEquals(100, counter1.registrationId());
+        assertEquals(CountersReader.RECORD_ALLOCATED, clientA.countersReader().getCounterState(counter1.id()));
+        assertEquals(counter1.registrationId(), clientA.countersReader().getCounterRegistrationId(counter1.id()));
+        assertEquals(NULL_VALUE, clientA.countersReader().getCounterOwnerId(counter1.id()));
+        assertEquals(COUNTER_TYPE_ID, clientA.countersReader().getCounterTypeId(counter1.id()));
+
+        final Counter counter2 = clientB.addStaticCounter(COUNTER_TYPE_ID, "test static counter", 200);
+
+        assertFalse(counter2.isClosed());
+        assertEquals(200, counter2.registrationId());
+        assertEquals(CountersReader.RECORD_ALLOCATED, clientB.countersReader().getCounterState(counter2.id()));
+        assertEquals(counter2.registrationId(), clientB.countersReader().getCounterRegistrationId(counter2.id()));
+        assertEquals(NULL_VALUE, clientB.countersReader().getCounterOwnerId(counter2.id()));
+        assertEquals("test static counter", clientB.countersReader().getCounterLabel(counter2.id()));
+        assertEquals(COUNTER_TYPE_ID, clientB.countersReader().getCounterTypeId(counter2.id()));
+
+        verify(availableCounterHandlerClientA, Mockito.after(1000L).never())
+            .onAvailableCounter(any(CountersReader.class), eq(counter1.registrationId()), eq(counter1.id()));
+        verify(availableCounterHandlerClientA, never())
+            .onAvailableCounter(any(CountersReader.class), eq(counter2.registrationId()), eq(counter2.id()));
+        verify(availableCounterHandlerClientB, never())
+            .onAvailableCounter(any(CountersReader.class), eq(counter1.registrationId()), eq(counter1.id()));
+        verify(availableCounterHandlerClientB, never())
+            .onAvailableCounter(any(CountersReader.class), eq(counter2.registrationId()), eq(counter2.id()));
+    }
+
+    @Test
+    @InterruptAfter(10)
+    void shouldReturnExistingStaticCounterAndNotUpdateAnything()
+    {
+        final AvailableCounterHandler availableCounterHandlerClientA = mock(AvailableCounterHandler.class);
+        final UnavailableCounterHandler unavailableCounterHandlerClientA = mock(UnavailableCounterHandler.class);
+        clientA.addAvailableCounterHandler(availableCounterHandlerClientA);
+        clientA.addUnavailableCounterHandler(unavailableCounterHandlerClientA);
+
+        final AvailableCounterHandler availableCounterHandlerClientB = mock(AvailableCounterHandler.class);
+        final UnavailableCounterHandler unavailableCounterHandlerClientB = mock(UnavailableCounterHandler.class);
+        clientB.addAvailableCounterHandler(availableCounterHandlerClientB);
+        clientB.addUnavailableCounterHandler(unavailableCounterHandlerClientB);
+
+        final long registrationId = 888;
+        ThreadLocalRandom.current().nextBytes(keyBuffer.byteArray());
+        final byte[] expectedKeyBytes = Arrays.copyOf(keyBuffer.byteArray(), keyBuffer.capacity());
+        final Counter counter1 = clientA.addStaticCounter(
+            COUNTER_TYPE_ID,
+            keyBuffer,
+            0,
+            keyBuffer.capacity(),
+            labelBuffer,
+            0,
+            COUNTER_LABEL.length(),
+            registrationId);
+
+        assertFalse(counter1.isClosed());
+        assertEquals(registrationId, counter1.registrationId());
+        assertEquals(CountersReader.RECORD_ALLOCATED, clientA.countersReader().getCounterState(counter1.id()));
+        assertEquals(counter1.registrationId(), clientA.countersReader().getCounterRegistrationId(counter1.id()));
+        assertEquals(NULL_VALUE, clientA.countersReader().getCounterOwnerId(counter1.id()));
+        assertEquals(COUNTER_TYPE_ID, clientA.countersReader().getCounterTypeId(counter1.id()));
+        assertEquals(COUNTER_LABEL, clientA.countersReader().getCounterLabel(counter1.id()));
+
+        final Counter counter2 = clientB.addStaticCounter(COUNTER_TYPE_ID, "test static counter", registrationId);
+
+        assertEquals(counter1.id(), counter2.id());
+        assertEquals(registrationId, counter2.registrationId());
+        assertEquals(CountersReader.RECORD_ALLOCATED, clientB.countersReader().getCounterState(counter2.id()));
+        assertEquals(registrationId, clientB.countersReader().getCounterRegistrationId(counter2.id()));
+        assertEquals(NULL_VALUE, clientB.countersReader().getCounterOwnerId(counter2.id()));
+        assertEquals(COUNTER_TYPE_ID, clientB.countersReader().getCounterTypeId(counter2.id()));
+        assertEquals(COUNTER_LABEL, clientB.countersReader().getCounterLabel(counter2.id()));
+
+        final MutableBoolean keyChecked = new MutableBoolean(false);
+        clientB.countersReader().forEach((counterId, typeId, keyBuffer, label) ->
+        {
+            if (counterId == counter1.id())
+            {
+                final byte[] actualKeyBytes = new byte[expectedKeyBytes.length];
+                keyBuffer.getBytes(0, actualKeyBytes, 0, expectedKeyBytes.length);
+                assertArrayEquals(expectedKeyBytes, actualKeyBytes);
+                keyChecked.set(true);
+            }
+        });
+        assertTrue(keyChecked.get());
+
+        verify(availableCounterHandlerClientA, Mockito.after(1000L).never())
+            .onAvailableCounter(any(CountersReader.class), eq(registrationId), eq(counter1.id()));
+        verify(availableCounterHandlerClientB, never())
+            .onAvailableCounter(any(CountersReader.class), eq(registrationId), eq(counter2.id()));
+    }
+
+    @Test
+    @InterruptAfter(10)
+    void shouldNotDeleteStaticCounterIfClosed()
+    {
+        final AvailableCounterHandler availableCounterHandlerClientA = mock(AvailableCounterHandler.class);
+        final UnavailableCounterHandler unavailableCounterHandlerClientA = mock(UnavailableCounterHandler.class);
+        clientA.addAvailableCounterHandler(availableCounterHandlerClientA);
+        clientA.addUnavailableCounterHandler(unavailableCounterHandlerClientA);
+
+        final AvailableCounterHandler availableCounterHandlerClientB = mock(AvailableCounterHandler.class);
+        final UnavailableCounterHandler unavailableCounterHandlerClientB = mock(UnavailableCounterHandler.class);
+        clientB.addAvailableCounterHandler(availableCounterHandlerClientB);
+        clientB.addUnavailableCounterHandler(unavailableCounterHandlerClientB);
+
+        final Counter counter1 = clientA.addStaticCounter(
+            COUNTER_TYPE_ID,
+            keyBuffer,
+            0,
+            keyBuffer.capacity(),
+            labelBuffer,
+            0,
+            COUNTER_LABEL.length(),
+            100);
+
+        assertFalse(counter1.isClosed());
+        assertEquals(100, counter1.registrationId());
+        assertEquals(CountersReader.RECORD_ALLOCATED, clientA.countersReader().getCounterState(counter1.id()));
+        assertEquals(counter1.registrationId(), clientA.countersReader().getCounterRegistrationId(counter1.id()));
+        assertEquals(NULL_VALUE, clientA.countersReader().getCounterOwnerId(counter1.id()));
+        assertEquals(COUNTER_TYPE_ID, clientB.countersReader().getCounterTypeId(counter1.id()));
+
+        counter1.close();
+        assertTrue(counter1.isClosed());
+
+        verify(unavailableCounterHandlerClientA, Mockito.after(1000L).never())
+            .onUnavailableCounter(any(CountersReader.class), eq(counter1.registrationId()), eq(counter1.id()));
+
+        assertEquals(CountersReader.RECORD_ALLOCATED, clientA.countersReader().getCounterState(counter1.id()));
+        assertEquals(CountersReader.RECORD_ALLOCATED, clientB.countersReader().getCounterState(counter1.id()));
+    }
+
+    @Test
+    @InterruptAfter(10)
+    @SuppressWarnings("indentation")
+    void shouldReturnErrorIfANonStaticCounterExistsForTypeIdRegistrationId()
+    {
+        final Counter counter = clientA.addCounter(COUNTER_TYPE_ID, "test session-specific counter");
+        assertNotEquals(NULL_VALUE, counter.registrationId());
+        assertEquals(clientA.clientId(), clientA.countersReader().getCounterOwnerId(counter.id()));
+
+        final RegistrationException registrationException = assertThrowsExactly(
+            RegistrationException.class,
+            () -> clientA.addStaticCounter(
+                COUNTER_TYPE_ID,
+                keyBuffer,
+                0,
+                keyBuffer.capacity(),
+                labelBuffer,
+                0,
+                COUNTER_LABEL.length(),
+                counter.registrationId()));
+
+        assertThat(
+            registrationException.getMessage(),
+            allOf(
+                containsString("cannot add static counter, because a non-static counter exists"),
+                containsString("counterId=" + counter.id()),
+                containsString("typeId=" + COUNTER_TYPE_ID),
+                containsString("registrationId=" + counter.registrationId()),
+                containsString("errorCodeValue=11")));
+    }
+
+    @Test
+    @InterruptAfter(10)
+    void shouldNotCloseStaticCounterWhenClientInstanceIsClosed()
+    {
+        final AtomicBoolean clientClosed = new AtomicBoolean();
+        clientA.addCloseHandler(() -> clientClosed.set(true));
+
+        final Counter counter1 = clientA.addStaticCounter(
+            COUNTER_TYPE_ID,
+            keyBuffer,
+            0,
+            keyBuffer.capacity(),
+            labelBuffer,
+            0,
+            COUNTER_LABEL.length(),
+            1);
+
+        final Counter counter2 = clientB.addStaticCounter(COUNTER_TYPE_ID + 2, "test static counter", 22);
+
+        final Counter counter3 = clientA.addCounter(COUNTER_TYPE_ID, "delete me");
+
+        clientA.close();
+
+        Tests.await(clientClosed::get);
+        assertFalse(counter1.isClosed());
+        assertTrue(counter3.isClosed());
+
+        Tests.await(() -> CountersReader.RECORD_RECLAIMED == clientB.countersReader().getCounterState(counter3.id()));
+        assertEquals(CountersReader.RECORD_ALLOCATED, clientB.countersReader().getCounterState(counter1.id()));
+        assertEquals(CountersReader.RECORD_ALLOCATED, clientB.countersReader().getCounterState(counter2.id()));
+        assertFalse(counter2.isClosed());
+    }
+
+    @Test
+    @InterruptAfter(10)
+    void shouldNotCloseStaticCounterIfClientTimesOut()
+    {
+        final AvailableCounterHandler availableCounterHandler = mock(AvailableCounterHandler.class);
+        final UnavailableCounterHandler unavailableCounterHandler = mock(UnavailableCounterHandler.class);
+        final AtomicReference<Throwable> error = new AtomicReference<>();
+        try (Aeron aeron = Aeron.connect(new Aeron.Context()
+            .aeronDirectoryName(driver.aeronDirectoryName())
+            .useConductorAgentInvoker(true)
+            .availableCounterHandler(availableCounterHandler)
+            .unavailableCounterHandler(unavailableCounterHandler)
+            .errorHandler(error::set)))
+        {
+            final AgentInvoker conductorAgentInvoker = aeron.conductorAgentInvoker();
+            assertNotNull(conductorAgentInvoker);
+            final CountersReader countersReader = clientA.countersReader();
+            assertEquals(0, countersReader.getCounterValue(SystemCounterDescriptor.CLIENT_TIMEOUTS.id()));
+
+            final Counter counter = aeron.addStaticCounter(COUNTER_TYPE_ID, "test", 42);
+            assertNotNull(counter);
+            assertFalse(counter.isClosed());
+            assertEquals(CountersReader.RECORD_ALLOCATED, aeron.countersReader().getCounterState(counter.id()));
+
+            final Counter counter2 = aeron.addCounter(COUNTER_TYPE_ID * 2, "delete me");
+
+            conductorAgentInvoker.invoke();
+
+            Tests.await(() -> 1 == countersReader.getCounterValue(SystemCounterDescriptor.CLIENT_TIMEOUTS.id()));
+
+            while (null == error.get())
+            {
+                conductorAgentInvoker.invoke();
+                Thread.yield();
+            }
+            final Throwable timeoutException = error.get();
+            if (timeoutException instanceof ClientTimeoutException)
+            {
+                assertEquals("FATAL - client timeout from driver", timeoutException.getMessage());
+            }
+            else if (timeoutException instanceof ConductorServiceTimeoutException)
+            {
+                assertThat(timeoutException.getMessage(), CoreMatchers.startsWith("FATAL - service interval exceeded"));
+            }
+            else if (timeoutException instanceof AeronException)
+            {
+                assertThat(
+                    timeoutException.getMessage(),
+                    CoreMatchers.startsWith("ERROR - unexpected close of heartbeat timestamp counter:"));
+            }
+            else
+            {
+                // on unknown error print stack trace
+                timeoutException.printStackTrace();
+            }
+
+            assertTrue(counter2.isClosed());
+            assertEquals(CountersReader.RECORD_RECLAIMED, aeron.countersReader().getCounterState(counter2.id()));
+            verify(availableCounterHandler).onAvailableCounter(
+                any(CountersReader.class), eq(counter2.registrationId()), eq(counter2.id()));
+            verify(unavailableCounterHandler).onUnavailableCounter(
+                any(CountersReader.class), eq(counter2.registrationId()), eq(counter2.id()));
+
+            assertFalse(counter.isClosed());
+            assertEquals(CountersReader.RECORD_ALLOCATED, aeron.countersReader().getCounterState(counter.id()));
+            verify(availableCounterHandler, never()).onAvailableCounter(
+                any(CountersReader.class), eq(counter.registrationId()), eq(counter.id()));
+            verify(unavailableCounterHandler, never()).onUnavailableCounter(
+                any(CountersReader.class), eq(counter.registrationId()), eq(counter.id()));
         }
     }
